refactor(core) removes circular dependancy
