+++ packages/core/src/api/config/generate-permissions.ts
@@ -2,8 +2,8 @@ import { stitchSchemas } from '@graphql-tools/stitch';
 import { GraphQLEnumType, GraphQLInputObjectType, GraphQLSchema } from 'graphql';
 import { GraphQLEnumValueConfigMap } from 'graphql/type/definition';
 
-import { DEFAULT_PERMISSIONS } from '../../common/constants';
-import { getAllPermissionsMetadata, PermissionDefinition } from '../../common/permission-definition';
+import { getAllPermissionsMetadata } from '../../common/constants';
+import { PermissionDefinition } from '../../common/permission-definition';
 
 const PERMISSION_DESCRIPTION = `@description
 Permissions for administrators and customers. Used to control access to
